perf, refactor: 細かいチューニング

+++ include/types.hpp
@@ -167,11 +167,11 @@ private:
          * フィールド上の得点を計算するメソッド
          */
         // 自分のパネルが置かれている部分の合計得点
-        i64 calc_mypanels_score();
+        i16 calc_mypanels_score();
         // 敵のパネルが置かれている部分の合計得点
-        i64 calc_enemypanels_score();
+        i16 calc_enemypanels_score();
         // 自分の合計と敵の合計の差。上記２つの関数を使って差を求めるよりも高速
-        i64 calc_sumpanel_score();
+        i16 calc_sumpanel_score();
 
 /*
  * make_atメソッド フィールド情報のアクセサメソッド
